Moved print_dp_or_event to formatting.py
